+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Fixed
 
 - Fix JSON template previews not rendering unpacked values correctly [#627](https://github.com/LucasPickering/slumber/issues/627)
+- Fix cycling backward (shift-tab) from the sidebar to the bottom pane
 
 ## [5.1.0] - 2026-02-20